Moved routes to the JG folder
